struct RdbPos moved to Common; some dependencies chagned
