+++ Common/RdbPos.h
@@ -1,0 +1,8 @@
+#pragma once
+class Project;
+
+struct RdbPos
+{
+	Project* rdb = nullptr;
+	int i_rec = 0;
+};
