+++ Prolog/ReadProlog.h
@@ -1,6 +1,7 @@
 #pragma once
 #include <string>
 #include "../Core/Project.h"
+#include "../Common/RdbPos.h"
 
 class ReadProlog
 {
